Split native waveform DTO modules
