+++ src/app_core/actions/native_shell_dtos.rs
@@ -12,13 +12,12 @@ use radiant::gui::list;
 use radiant::gui::panel;
 use radiant::gui::range;
 use radiant::gui::retained;
-use radiant::gui::types::ImageRgba;
 use radiant::gui::visualization;
-use std::sync::Arc;
 
 mod automation;
 mod browser;
 mod retained_segments;
+mod waveform;
 
 pub use self::automation::{
     AutomationBounds, AutomationNodeId, AutomationNodeSnapshot, AutomationRole,
@@ -30,6 +29,13 @@ pub use self::browser::{
     PlaybackAgeBucket, PlaybackAgeFilterChip,
 };
 pub use self::retained_segments::{DirtySegments, SegmentRevisions};
+pub use self::waveform::{
+    WaveformChannelViewModel, WaveformChromeModel, WaveformChromeStateModel,
+    WaveformEditPreviewModel, WaveformFeedbackEventsModel, WaveformImagePreviewModel,
+    WaveformMotionModel, WaveformPanelModel, WaveformPresentationModel, WaveformSlicePreviewModel,
+    WaveformSurfaceModel, WaveformToolStateModel, WaveformTransportModel, WaveformViewportModel,
+    parse_waveform_tempo_number_text,
+};
 
 /// Shared storage used by retained app-model snapshots.
 pub type RetainedVec<T> = retained::RetainedVec<T>;
@@ -98,116 +104,6 @@ pub type MapRenderModeModel = visualization::PointRenderMode;
 
 /// Summary of map state consumed by the native shell map tab.
 pub type MapPanelModel = visualization::SpatialPanel;
-
-/// Channel-view mode used by waveform rendering.
-pub type WaveformChannelViewModel = visualization::ChannelViewMode;
-
-/// Normalized waveform viewport state exposed to the native shell.
-pub type WaveformViewportModel = visualization::TimelineViewport;
-
-/// Waveform cursor, playhead, and selection transport state exposed to the native shell.
-pub type WaveformTransportModel = visualization::TimelineTransportState;
-
-/// Waveform edit selection and fade-preview state exposed to the native shell.
-pub type WaveformEditPreviewModel = visualization::TimelineEditPreview;
-
-/// One detected Wavecrate waveform slice preview exposed to the native shell.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct WaveformSlicePreviewModel {
-    /// Slice range in normalized waveform precision.
-    pub range: NormalizedRangeModel,
-    /// Whether this slice is currently selected for edit operations.
-    pub selected: bool,
-    /// Whether this slice is focused for keyboard review.
-    pub focused: bool,
-    /// Whether this slice is marked for sample export.
-    pub marked_for_export: bool,
-    /// Whether this slice belongs to the duplicate-cleanup candidate batch.
-    pub review_candidate: bool,
-    /// Whether this slice is currently exempted from duplicate cleanup.
-    pub review_exempted: bool,
-}
-
-/// One-shot waveform feedback event tokens exposed to the native shell.
-pub type WaveformFeedbackEventsModel = visualization::TimelineFeedbackEvents;
-
-/// Waveform guide/repeat/label presentation state exposed to the native shell.
-pub type WaveformPresentationModel = visualization::TimelinePresentationState;
-
-/// Retained waveform raster preview state exposed to the native shell.
-pub type WaveformImagePreviewModel = visualization::SignalRasterPreview;
-
-/// Waveform display chrome state exposed to the native shell.
-pub type WaveformChromeStateModel = visualization::SignalChromeState;
-
-/// Waveform tool availability state exposed to the native shell.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct WaveformToolStateModel {
-    /// Whether loop playback is locked against sample-driven updates.
-    pub lock_enabled: bool,
-    /// Whether normalized audition playback is enabled.
-    pub audition_enabled: bool,
-    /// Whether BPM snapping is enabled.
-    pub primary_snap_enabled: bool,
-    /// Whether playback BPM grids and snapping use selection-relative anchors.
-    pub relative_grid_enabled: bool,
-    /// Whether transient snapping is enabled.
-    pub secondary_snap_enabled: bool,
-    /// Whether transient markers are visible.
-    pub markers_visible: bool,
-    /// Whether slice review mode is active.
-    pub review_mode_enabled: bool,
-    /// Whether exact-duplicate cleanup can be applied from the waveform toolbar.
-    pub cleanup_available: bool,
-}
-
-impl Default for WaveformToolStateModel {
-    fn default() -> Self {
-        Self {
-            lock_enabled: false,
-            audition_enabled: false,
-            primary_snap_enabled: false,
-            relative_grid_enabled: false,
-            secondary_snap_enabled: false,
-            markers_visible: true,
-            review_mode_enabled: false,
-            cleanup_available: false,
-        }
-    }
-}
-
-impl WaveformToolStateModel {
-    /// Build waveform tool state from explicit Wavecrate workflow flags.
-    #[allow(clippy::too_many_arguments)]
-    pub fn new(
-        lock_enabled: bool,
-        audition_enabled: bool,
-        primary_snap_enabled: bool,
-        relative_grid_enabled: bool,
-        secondary_snap_enabled: bool,
-        markers_visible: bool,
-        review_mode_enabled: bool,
-        cleanup_available: bool,
-    ) -> Self {
-        Self {
-            lock_enabled,
-            audition_enabled,
-            primary_snap_enabled,
-            relative_grid_enabled,
-            secondary_snap_enabled,
-            markers_visible,
-            review_mode_enabled,
-            cleanup_available,
-        }
-    }
-}
-
-/// Aggregated waveform timeline surface state exposed to the native shell.
-pub type WaveformSurfaceModel = visualization::TimelineSurfaceState<WaveformSlicePreviewModel>;
-
-/// Aggregated waveform motion state exposed to the native shell.
-pub type WaveformMotionModel =
-    visualization::TimelineMotionState<WaveformSlicePreviewModel, WaveformToolStateModel>;
 
 /// Render data for one point shown in the native map canvas.
 pub type MapPointModel = visualization::SpatialPoint;
@@ -722,295 +618,6 @@ pub enum ConfirmPromptKind {
     PurgeRetainedFolderDeletes,
     /// Pending options-panel default-identifier prompt.
     OptionsDefaultIdentifier,
-}
-
-/// Waveform preview metadata consumed by the native shell.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct WaveformPanelModel {
-    /// Display label for the loaded sample, when any.
-    pub loaded_label: Option<String>,
-    /// Whether a newly focused sample is still loading waveform data.
-    pub loading: bool,
-    /// Whether a replacement waveform image is still rendering in the background.
-    pub image_rendering: bool,
-    /// Cursor position in normalized milli-units.
-    pub cursor_milli: Option<u16>,
-    /// Playhead position in normalized milli-units.
-    pub playhead_milli: Option<u16>,
-    /// Playhead position in normalized micro-units (`0..=1_000_000`).
-    pub playhead_micros: Option<u32>,
-    /// Current waveform selection bounds.
-    pub selection_milli: Option<NormalizedRangeModel>,
-    /// Preview slices detected from silence-splitting the loaded waveform.
-    pub slices: Vec<WaveformSlicePreviewModel>,
-    /// One-shot token incremented when a waveform-selection export is queued.
-    pub selection_export_flash_nonce: u64,
-    /// One-shot token incremented when a queued waveform-selection export fails.
-    pub selection_export_failure_flash_nonce: u64,
-    /// One-shot token incremented when preview edit fades are committed.
-    pub edit_selection_apply_flash_nonce: u64,
-    /// Current waveform edit-selection bounds.
-    pub edit_selection_milli: Option<NormalizedRangeModel>,
-    /// End position for the edit fade-in region in normalized milli-units.
-    pub edit_fade_in_end_milli: Option<u16>,
-    /// End position for the edit fade-in region in normalized micro-units.
-    pub edit_fade_in_end_micros: Option<u32>,
-    /// Start position for the edit fade-in mute region in normalized milli-units.
-    pub edit_fade_in_mute_start_milli: Option<u16>,
-    /// Start position for the edit fade-in mute region in normalized micro-units.
-    pub edit_fade_in_mute_start_micros: Option<u32>,
-    /// Fade-in curve tension in normalized milli-units (`0..=1000`).
-    pub edit_fade_in_curve_milli: Option<u16>,
-    /// Start position for the edit fade-out region in normalized milli-units.
-    pub edit_fade_out_start_milli: Option<u16>,
-    /// Start position for the edit fade-out region in normalized micro-units.
-    pub edit_fade_out_start_micros: Option<u32>,
-    /// End position for the edit fade-out mute region in normalized milli-units.
-    pub edit_fade_out_mute_end_milli: Option<u16>,
-    /// End position for the edit fade-out mute region in normalized micro-units.
-    pub edit_fade_out_mute_end_micros: Option<u32>,
-    /// Fade-out curve tension in normalized milli-units (`0..=1000`).
-    pub edit_fade_out_curve_milli: Option<u16>,
-    /// Visible view start in normalized milli-units.
-    pub view_start_milli: u16,
-    /// Visible view end in normalized milli-units.
-    pub view_end_milli: u16,
-    /// Visible view start in normalized micro-units (`0..=1_000_000`).
-    pub view_start_micros: u32,
-    /// Visible view end in normalized micro-units (`0..=1_000_000`).
-    pub view_end_micros: u32,
-    /// Visible view start in normalized nanounits (`0..=1_000_000_000`).
-    pub view_start_nanos: u32,
-    /// Visible view end in normalized nanounits (`0..=1_000_000_000`).
-    pub view_end_nanos: u32,
-    /// Quarter-note beat spacing in normalized micro-units when BPM/grid data is available.
-    pub beat_step_micros: Option<u32>,
-    /// BPM grid origin in normalized micro-units.
-    pub bpm_grid_origin_micros: u32,
-    /// Whether loop playback is enabled.
-    pub loop_enabled: bool,
-    /// Optional tempo label rendered in waveform metadata.
-    pub tempo_label: Option<String>,
-    /// Optional zoom label rendered in waveform metadata.
-    pub zoom_label: Option<String>,
-    /// Cached signature for waveform image updates.
-    pub waveform_image_signature: Option<u64>,
-    /// Optional rasterized waveform payload for rendering the waveform preview.
-    pub waveform_image: Option<Arc<ImageRgba>>,
-}
-
-impl Default for WaveformPanelModel {
-    fn default() -> Self {
-        Self {
-            loaded_label: None,
-            loading: false,
-            image_rendering: false,
-            cursor_milli: None,
-            playhead_milli: None,
-            playhead_micros: None,
-            selection_milli: None,
-            slices: Vec::new(),
-            selection_export_flash_nonce: 0,
-            selection_export_failure_flash_nonce: 0,
-            edit_selection_apply_flash_nonce: 0,
-            edit_selection_milli: None,
-            edit_fade_in_end_milli: None,
-            edit_fade_in_end_micros: None,
-            edit_fade_in_mute_start_milli: None,
-            edit_fade_in_mute_start_micros: None,
-            edit_fade_in_curve_milli: None,
-            edit_fade_out_start_milli: None,
-            edit_fade_out_start_micros: None,
-            edit_fade_out_mute_end_milli: None,
-            edit_fade_out_mute_end_micros: None,
-            edit_fade_out_curve_milli: None,
-            view_start_milli: 0,
-            view_end_milli: 1000,
-            view_start_micros: 0,
-            view_end_micros: 1_000_000,
-            view_start_nanos: 0,
-            view_end_nanos: 1_000_000_000,
-            beat_step_micros: None,
-            bpm_grid_origin_micros: 0,
-            loop_enabled: false,
-            tempo_label: None,
-            zoom_label: None,
-            waveform_image_signature: None,
-            waveform_image: None,
-        }
-    }
-}
-
-impl WaveformPanelModel {
-    /// Return this panel's generic normalized timeline viewport.
-    pub fn viewport(&self) -> WaveformViewportModel {
-        WaveformViewportModel::new(
-            self.view_start_milli,
-            self.view_end_milli,
-            self.view_start_micros,
-            self.view_end_micros,
-            self.view_start_nanos,
-            self.view_end_nanos,
-        )
-    }
-
-    /// Return this panel's generic timeline transport state.
-    pub fn transport(&self) -> WaveformTransportModel {
-        WaveformTransportModel::new(
-            self.cursor_milli,
-            self.playhead_milli,
-            self.playhead_micros,
-            self.selection_milli,
-        )
-    }
-
-    /// Return this panel's generic timeline edit preview.
-    pub fn edit_preview(&self) -> WaveformEditPreviewModel {
-        WaveformEditPreviewModel::from_parts(visualization::TimelineEditPreviewParts {
-            selection: self.edit_selection_milli,
-            leading_end_milli: self.edit_fade_in_end_milli,
-            leading_end_micros: self.edit_fade_in_end_micros,
-            leading_inner_start_milli: self.edit_fade_in_mute_start_milli,
-            leading_inner_start_micros: self.edit_fade_in_mute_start_micros,
-            leading_curve_milli: self.edit_fade_in_curve_milli,
-            trailing_start_milli: self.edit_fade_out_start_milli,
-            trailing_start_micros: self.edit_fade_out_start_micros,
-            trailing_inner_end_milli: self.edit_fade_out_mute_end_milli,
-            trailing_inner_end_micros: self.edit_fade_out_mute_end_micros,
-            trailing_curve_milli: self.edit_fade_out_curve_milli,
-        })
-    }
-
-    /// Return this panel's generic timeline feedback events.
-    pub fn feedback_events(&self) -> WaveformFeedbackEventsModel {
-        WaveformFeedbackEventsModel::new(
-            self.selection_export_flash_nonce,
-            self.selection_export_failure_flash_nonce,
-            self.edit_selection_apply_flash_nonce,
-        )
-    }
-
-    /// Return this panel's generic timeline presentation state.
-    pub fn presentation(&self) -> WaveformPresentationModel {
-        WaveformPresentationModel::new(
-            self.beat_step_micros,
-            self.bpm_grid_origin_micros,
-            self.loop_enabled,
-            self.tempo_label.clone(),
-            self.zoom_label.clone(),
-        )
-    }
-
-    /// Return this panel's generic retained raster preview.
-    pub fn image_preview(&self) -> WaveformImagePreviewModel {
-        WaveformImagePreviewModel::new(
-            self.loaded_label.clone(),
-            self.loading,
-            self.image_rendering,
-            self.waveform_image_signature,
-            self.waveform_image.clone(),
-        )
-    }
-
-    /// Return this panel's generic normalized timeline surface state.
-    pub fn timeline_surface(&self) -> WaveformSurfaceModel {
-        WaveformSurfaceModel::from_parts(visualization::TimelineSurfaceParts {
-            viewport: self.viewport(),
-            transport: self.transport(),
-            edit_preview: self.edit_preview(),
-            feedback_events: self.feedback_events(),
-            presentation: self.presentation(),
-            raster_preview: self.image_preview(),
-            markers: self.slices.clone(),
-        })
-    }
-}
-
-/// Waveform chrome copy used by metadata lines and control surfaces.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct WaveformChromeModel {
-    /// Extra transport metadata hint shown alongside waveform labels.
-    pub transport_hint: String,
-    /// Whether compare-anchor replay is currently available.
-    pub compare_anchor_available: bool,
-    /// Label for the stored compare anchor, when available.
-    pub compare_anchor_label: Option<String>,
-    /// Whether loop state is locked against sample-driven auto-updates.
-    pub loop_lock_enabled: bool,
-    /// Current channel-view mode used by waveform rendering.
-    pub channel_view: WaveformChannelViewModel,
-    /// Whether normalized audition playback is enabled.
-    pub normalized_audition_enabled: bool,
-    /// Whether BPM snapping is enabled for waveform edits.
-    pub bpm_snap_enabled: bool,
-    /// Whether playback BPM grids and snapping use selection-relative anchors.
-    pub relative_bpm_grid_enabled: bool,
-    /// Whether transient snapping is enabled for waveform edits.
-    pub transient_snap_enabled: bool,
-    /// Whether transient markers are visible on the waveform.
-    pub transient_markers_enabled: bool,
-    /// Whether slice mode is currently active.
-    pub slice_mode_enabled: bool,
-    /// Whether the current slice batch is an exact-duplicate cleanup preview.
-    pub exact_duplicate_cleanup_available: bool,
-}
-
-impl Default for WaveformChromeModel {
-    fn default() -> Self {
-        Self {
-            transport_hint: String::from("transport idle"),
-            compare_anchor_available: false,
-            compare_anchor_label: None,
-            loop_lock_enabled: false,
-            channel_view: WaveformChannelViewModel::Mono,
-            normalized_audition_enabled: false,
-            bpm_snap_enabled: false,
-            relative_bpm_grid_enabled: false,
-            transient_snap_enabled: false,
-            transient_markers_enabled: true,
-            slice_mode_enabled: false,
-            exact_duplicate_cleanup_available: false,
-        }
-    }
-}
-
-impl WaveformChromeModel {
-    /// Return this chrome model's generic signal visualization display state.
-    pub fn signal_chrome(&self) -> WaveformChromeStateModel {
-        WaveformChromeStateModel::new(
-            self.transport_hint.clone(),
-            self.compare_anchor_available,
-            self.compare_anchor_label.clone(),
-            self.channel_view,
-        )
-    }
-
-    /// Return this chrome model's generic signal visualization tool state.
-    pub fn signal_tools(&self) -> WaveformToolStateModel {
-        WaveformToolStateModel::new(
-            self.loop_lock_enabled,
-            self.normalized_audition_enabled,
-            self.bpm_snap_enabled,
-            self.relative_bpm_grid_enabled,
-            self.transient_snap_enabled,
-            self.transient_markers_enabled,
-            self.slice_mode_enabled,
-            self.exact_duplicate_cleanup_available,
-        )
-    }
-}
-
-/// Extract the numeric BPM portion from one projected tempo label.
-pub fn parse_waveform_tempo_number_text(label: &str) -> Option<String> {
-    let number = label.split_ascii_whitespace().next()?.trim();
-    if number.is_empty() {
-        return None;
-    }
-    let parsed = number.parse::<f32>().ok()?;
-    if !parsed.is_finite() || parsed <= 0.0 {
-        return None;
-    }
-    Some(number.to_string())
 }
 
 /// Logical focus buckets projected into the native runtime.
